refactor: MatesUiModel 네이밍 변경
MatesUiModel이 공통적으로 사용될 예정이므로 MemberUiModel로 변경

- 이전: MatesUiModel
- 이후: MemberUiModel
